v0.2.2.1
app.go: players' and grenades' positions are now saved with some frequency; view direction of a player is now saved as integer; both positions and view directions may be saved as deltas (although it doesn't seem to be efficient)

+++ app/data_desc.go
@@ -43,10 +43,10 @@ type GrenadePositionInfo struct {
 }
 
 type PlayerMovementInfo struct {
-	SteamID		int64		`bson:"SteamID"`
+	SteamID		int64			`bson:"SteamID"`
 	Position	Int16Vector3	`bson:"Position"`
-	ViewX		float32		`bson:"ViewX"`
-	ViewY		float32		`bson:"ViewY"`
+	ViewX		int16			`bson:"ViewX"`
+	ViewY		int16			`bson:"ViewY"`
 }
 
 type PlayerStaticInfo struct {
